Add files via upload
This is my undergraduate physics thesis I completed as the final assignment toward my Bachelor's in Physics. There is no abstract, but as an overview, the research I conducted focused on the perfect Laplace operator obtained through a Renormalization Group Transformation on the discrete 2D action for a free scalar field. The perfect Laplacian has zero discretization errors, so I wanted to compare it with common finite difference schemes on different Poisson systems to assess its efficacy.